+++ nix/hydra/shell.nix
@@ -77,6 +77,7 @@ let
       hydra-prelude
       hydra-cardano-api
       hydra-test-utils
+      hydra-plutus-extras
       plutus-cbor
       plutus-merkle-tree
       hydra-plutus